Implicit conversions not allowed in array init.
When compiling with gcc --std=c++0x, these are not allowed:
See 8.5.4/7: A narrowing conversion is an implicit conversion.
And http://gcc.gnu.org/bugzilla/show_bug.cgi?id=49793

+++ VtkVis/VtkVisHelper.cpp
@@ -27,7 +27,10 @@ vtkImageData* VtkVisHelper::QImageToVtkImageData(QImage &img)
 		for (size_t i = 0; i < imgWidth; i++)
 		{
 			QRgb pix = img.pixel(i,j);
-			const float color[3] = { qRed(pix), qGreen(pix), qBlue(pix) };
+			const float color[3] = { static_cast<float>(qRed(pix)),
+									 static_cast<float>(qGreen(pix)),
+									 static_cast<float>(qBlue(pix))
+								   };
 			data->SetTuple(j * imgWidth + i, color);
 		}
